requirements.txt
pyspark
boto3
pyodbc
pandas
pymongo
uvicorn
fastapi
pickle-mixin
python-dotenv
s3fs
requests